Fixed two bugs:  (1) aperm.factor was using the original dimension attribute in the output but is now fixed, (2) the two assignment functions `[[<-.data.list` and `$<-.data.list` now properly accept NULL as a value.
There are now summary.data.list and aperm.data.list methods. Adding the latter required the addition of a match.dnames attribute to data list objects.

There is also now a new variablize family of functions, which convert all but the first dimensions of its argument into 'variables' in a data frame.

git-svn-id: svn://scm.r-forge.r-project.org/svnroot/multitable@15 edb9625f-4e0d-4859-8d74-9fd3b1da38cb